feat : Maker update,delete 인증 추가

+++ src/main/java/com/isfive/usearth/domain/maker/service/MakerService.java
@@ -1,5 +1,7 @@
 package com.isfive.usearth.domain.maker.service;
 
+import com.isfive.usearth.exception.BusinessException;
+import com.isfive.usearth.exception.ErrorCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -49,15 +51,19 @@ public class MakerService {
 	}
 
 	@Transactional
-	public MakerResponse updateMakerById(Long id, MakerUpdate makerUpdate) {
+	public MakerResponse updateMakerById(String username, Long id, MakerUpdate makerUpdate) {
 		Maker maker = makerRepository.findByIdOrThrow(id);
+		if (!maker.getName().equals(username))
+			throw new BusinessException(ErrorCode.AUTHENTICATION_FAILED);
 		maker.update(makerUpdate);
 		return MakerResponse.fromEntity(makerRepository.save(maker));
 	}
 
 	@Transactional
-	public void removeMakerById(Long id) {
+	public void removeMakerById(String username, Long id) {
 		Maker maker = makerRepository.findByIdOrThrow(id);
+		if (!maker.getName().equals(username))
+			throw new BusinessException(ErrorCode.AUTHENTICATION_FAILED);
 		maker.delete();
 	}
 }
